Restructure prompt and instructions into their own dir
